requirements.txt
gevent==25.5.1
greenlet==3.2.3
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
setuptools==80.9.0
websockets==15.0.1
zope.event==5.1
zope.interface==7.2
